refactor(kassensystem): architecture change refactor(security): refactored security aspects to commons. more consistent and overall improvements to AuthContext and Jwt Authentication / Filters refactor(jwt): added static jwt builders, payloads and enums. this allows for more consistent reading and building of them refactor(vaadin): extracted abstract ui components to new commons-ui module feat(lock): implemented account locking feat(lock): implemented account specific login. admin cant login to kassensystem and cashier not to fillialensystem feat(configuration): added autoconfiguration classes to commons module. this allows common filters, values, beans / components to only be set once and be implemented into using modules feat(vaadin): added global errorHandler overlay. on occurring exception, the errorHandler overlay outputs the exception including the stacktrace, message, title and other oddities, which can be more easily copied / be made aware of feat(utility): added many helper utility methods which allow for more consistent operations. AuthContext, AuthContextHolder, DateTimeFormat, AuthWhitelist ... bugfix(dto): fixed null value in AccountDTO bugfix(pom): improved pom dependency versioning. BOMs and properties properly out layered to parents

+++ filialensystem/src/main/java/de/fhdw/fillialensystem/api/controller/ReceiptController.java
@@ -2,25 +2,20 @@ package de.fhdw.fillialensystem.api.controller;
 
 import de.fhdw.commons.api.dto.ReceiptDTO;
 import de.fhdw.commons.api.dto.ReceiptLinkArticleDTO;
-import de.fhdw.commons.utility.AuthContext;
+import de.fhdw.commons.security.utility.security.auth.AuthContext;
 import de.fhdw.fillialensystem.api.mapper.ReceiptLinkArticleMapper;
 import de.fhdw.fillialensystem.api.mapper.ReceiptMapper;
-import de.fhdw.fillialensystem.persistence.entity.Store;
-import de.fhdw.fillialensystem.persistence.entity.StoreLinkStock;
-import de.fhdw.fillialensystem.persistence.entity.imported.Article;
 import de.fhdw.fillialensystem.persistence.service.ReceiptService;
 import de.fhdw.fillialensystem.persistence.service.StoreLinkStockService;
 import de.fhdw.fillialensystem.persistence.service.StoreService;
 import de.fhdw.fillialensystem.persistence.service.imported.ArticleService;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.persistence.EntityExistsException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/receipt")
@@ -48,7 +43,7 @@ public class ReceiptController {
         return ResponseEntity
                 .status(HttpStatus.CREATED)
                 .body(receiptMapper.toDto(
-                        receiptService.createReceipt(authContext.getRegisterId().longValue(), authContext.getUuid(), receiptArticles
+                        receiptService.createReceipt(authContext.getRegisterId(), authContext.getUuid(), receiptArticles
                                 .stream()
                                 .map(receiptLinkArticleMapper::toEntity)
                                 .toList())));
